#5 classes and creation composer
