+++ entity/User.php
@@ -1,4 +1,5 @@
 <?php
+namespace App\Entity;
 
 class User{
