+++ index.php
@@ -1,22 +1,19 @@
 <?php
-include_once('./Manager/manager.php');
-include_once('./Classe/Post.php');
-include_once('./Classe/User.php');
-include_once('./Classe/Comment.php');
-
+require 'vendor/autoload.php';
+include_once('./manager/manager.php');
 
 //var_dump($_GET);
 
 // ROUTER
 if (count($_GET)==0){
-    include_once('./Controller/homeController.php');
+    include_once('./controller/homeController.php');
     displayHome();
 }else{
     $controller=$_GET['controller'];
     $action=$_GET['action'];
 
     if ($controller == 'post'){
-        include_once('./Controller/postController.php');
+        include_once('./controller/postController.php');
         if ($action == 'displayList'){
             displayList();
         }elseif($action == 'displayOne'){
@@ -26,7 +23,7 @@ if (count($_GET)==0){
         }
     }
     if ($controller == 'home'){
-        include_once('./Controller/homeController.php');
+        include_once('./controller/homeController.php');
         if ($action == 'displayHome'){
             displayHome();
         }elseif($action == 'doSendEmail'){
@@ -34,7 +31,7 @@ if (count($_GET)==0){
         }
     }
     if ($controller == 'user'){
-        include_once('./Controller/userController.php');
+        include_once('./controller/userController.php');
         if ($action == 'displayLogin'){
             displayLogin();
         }elseif($action == 'displayRegister'){
@@ -50,4 +47,4 @@ if (count($_GET)==0){
 
 // Laisse ça commenté en verra tt à la fin car les variables $content ne sont pas
 // pour l'instant accessibles ici (pas de return)
-// include_once('./View/layout.html.php');
+// include_once('./view/layout.html.php');
